add to field in db to EmployeeOnboard

+++ src/services/EmployeeOnboard.service.js
@@ -1,18 +1,13 @@
 import prisma from "../config/db.js";
 import { BadRequestError } from "../utils/error.js";
-const createEmployeeOnboard = async (employeeId, data) => {
-  const empId = await prisma.employee.findFirst({
-    where: {
-      id: Number(employeeId),
-    },
-  });
-  if (!empId) {
-    throw new BadRequestError("Employee not found");
-  }
+
+const createEmployeeOnboard = async (data) => {
   return await prisma.employeeOnboard.create({
     data: {
-      ...data,
-      employeeId: Number(employeeId),
+      joiningDate: data.joiningDate,
+      candidateName: data.candidateName,
+      jobTitle: data.jobTitle,
+
       taskCheckLists: {
         create: [
           {
@@ -40,7 +35,6 @@ const createEmployeeOnboard = async (employeeId, data) => {
     },
     include: {
       taskCheckLists: true,
-      employee: true,
     },
   });
 };
@@ -48,13 +42,22 @@ const createEmployeeOnboard = async (employeeId, data) => {
 const fetchEmployeeOnboards = async () => {
   return await prisma.employeeOnboard.findMany({
     include: {
-      employee: true,
       taskCheckLists: true,
     },
   });
 };
 
 const updateEmployeeOnboard = async (id, data) => {
+  const onboard = await prisma.employeeOnboard.findUnique({
+    where: {
+      id: Number(id),
+    },
+  });
+
+  if (!onboard) {
+    throw new BadRequestError("Employee Onboard record not found");
+  }
+
   return await prisma.employeeOnboard.update({
     where: {
       id: Number(id),
@@ -64,6 +67,16 @@ const updateEmployeeOnboard = async (id, data) => {
 };
 
 const deleteEmployeeOnboard = async (id) => {
+  const onboard = await prisma.employeeOnboard.findUnique({
+    where: {
+      id: Number(id),
+    },
+  });
+
+  if (!onboard) {
+    throw new BadRequestError("Employee Onboard record not found");
+  }
+
   return await prisma.employeeOnboard.delete({
     where: {
       id: Number(id),
@@ -95,7 +108,7 @@ const updateTaskChecklist = async (id, data) => {
     }
   }
 
-  return await prisma.taskChecklist.create({
+  return await prisma.taskChecklist.update({
     where: {
       id: Number(id),
     },
@@ -107,14 +120,11 @@ const updateTaskChecklist = async (id, data) => {
 };
 
 const getEmployeeOnboardById = async (id) => {
-  console.log("id", id);
-
   const onboard = await prisma.employeeOnboard.findUnique({
     where: {
       id: Number(id),
     },
     include: {
-      employee: true,
       taskCheckLists: true,
     },
   });
